Setup capabilites only when no -N CLI option present.
Before  Util.hs has set capabilities always to the number of processors,
which works well only on a personal workstations with few processors.

On a cluster one needs to:
1) First check that -N option was not given.
2) Then limit number of capabilities to 12 at most.

Both things are now done so that it should work equally well on cluster
and on workstation with the same binary.